chore:refactor(tests): package oot_test_* + run_test.sh as standalone installable oot-framework under tests/oot_framework/
