+++ .github/workflows/model_ops_tests.yaml
@@ -9,8 +9,8 @@ on:
       - docs/**
       - README.md
   # Removing from CICD till the ops tests get matured
-  # pull_request:
-  #   branches: [main]
+  pull_request:
+    branches: [main]
   # merge_group:
   #   types: [checks_requested]
