+++ tests/oot_framework/oot_test_common_methods_invocations.py
