+++ tests/oot_framework/oot_test_constants.py
